Delete unnecessary files and add new files for store entities and repeated transactions
