+++ package.json
@@ -1,12 +1,13 @@
 {
   "name": "luca-ledger",
-  "version": "1.3.2",
+  "version": "2.0.0",
   "homepage": "https://jwaspin.github.io/luca-ledger/",
   "type": "module",
   "license": "MIT",
   "description": "A simple ledger app for tracking your expenses.",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite --port 8080",
+    "start": "vite --port 8080",
     "build": "vite build",
     "lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
@@ -20,6 +21,7 @@
     "@mui/system": "^5.14.15",
     "@mui/x-date-pickers": "^6.17.0",
     "@reduxjs/toolkit": "^1.9.7",
+    "ajv": "^8.12.0",
     "dayjs": "^1.11.10",
     "immer": "^10.0.3",
     "jszip": "^3.10.1",
@@ -27,8 +29,7 @@
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.3",
     "react-router-dom": "^6.17.0",
-    "uuid": "^9.0.1",
-    "yup": "^1.3.2"
+    "uuid": "^9.0.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.33",
