+++ src/schemas/categories.json
@@ -1,0 +1,37 @@
+{
+  "type": "object",
+  "properties": {
+    "id": {
+      "type": "string"
+    },
+    "name": {
+      "type": "string"
+    },
+    "description": {
+      "type": "string",
+      "nullable": true
+    },
+    "parentId": {
+      "type": "string",
+      "nullable": true
+    },
+    "defaultCategoryId": {
+      "type": "string",
+      "nullable": true
+    },
+    "categoryStatus": {
+      "type": "string",
+      "enum": ["DEFAULT", "MODIFIED", "CUSTOM"]
+    },
+    "createdAt": {
+      "type": "string",
+      "format": "date-time"
+    },
+    "updatedAt": {
+      "type": "string",
+      "format": "date-time",
+      "nullable": true
+    }
+  },
+  "required": ["id", "name", "createdAt"]
+}
