+++ src/schemas/entities.json
@@ -1,0 +1,14 @@
+{
+  "entities": {
+    "type": "object",
+    "properties": {
+      "id": { "type": "string" },
+      "name": { "type": "string" },
+      "description": { "type": "string", "nullable": true },
+      "typeId": { "type": "string" },
+      "createdAt": { "type": "string", "format": "date-time" },
+      "updatedAt": { "type": "string", "format": "date-time", "nullable": true }
+    },
+    "required": ["id", "name", "typeId", "createdAt"]
+  }
+}
