+++ src/schemas/repeatedTransactions.json
@@ -1,0 +1,17 @@
+"repeatedTransactions": {
+    "type": "object",
+    "properties": {
+      "id": {"type": "string"},
+      "payorId": {"type": "string"},
+      "payeeId": {"type": "string"},
+      "categoryId": {"type": "string"},
+      "amount": {"type": "number"},
+      "frequency": {"type": "string", "enum": ["day", "week", "month", "occurrence"]},
+      "interval": {"type": "integer"},
+      "description": {"type": "string"},
+      "state": {"type": "string", "enum": ["active", "paused", "completed"]},
+      "createdAt": {"type": "string", "format": "date-time"},
+      "updatedAt": {"type": "string", "format": "date-time", "nullable": true}
+    },
+    "required": ["id", "payorId", "payeeId", "categoryId", "amount", "frequency", "interval", "state", "createdAt"]
+  },
