+++ src/schemas/transactions.json
@@ -1,0 +1,16 @@
+{
+  "type": "object",
+  "properties": {
+    "id": {"type": "string"},
+    "payorId": {"type": "string"},
+    "payeeId": {"type": "string"},
+    "categoryId": {"type": "string"},
+    "amount": {"type": "number"},
+    "date": {"type": "string", "format": "date"},
+    "description": {"type": "string"},
+    "state": {"type": "string", "enum": ["planned", "scheduled", "pending", "completed", "cancelled"]},
+    "createdAt": {"type": "string", "format": "date-time"},
+    "updatedAt": {"type": "string", "format": "date-time", "nullable": true}
+  },
+  "required": ["id", "payorId", "payeeId", "categoryId", "amount", "date", "state", "createdAt"]
+}
