--- src/store/accounts/constants.js
@@ -1,12 +1,0 @@
-export const AccountType = Object.freeze({
-  SAVINGS: 'Savings',
-  CHECKING: 'Checking',
-  CREDIT_CARD: 'Credit Card',
-});
-
-export const AccountFields = Object.freeze({
-  ID: 'id',
-  NAME: 'name',
-  TYPE: 'type',
-  STATEMENT_DAY: 'statementDay',
-});
