--- src/store/accounts/selectors.js
@@ -1,4 +1,0 @@
-export const selectAccountById = (id) => (state) =>
-  state.accounts.find((account) => account.id === id);
-
-export const selectAccounts = (state) => state.accounts;
