--- src/store/categories/constants.js
@@ -1,54 +1,0 @@
-export const CategoryFields = Object.freeze({
-  ID: 'id',
-  NAME: 'name',
-  SUB_CATEGORIES: 'subCategories',
-});
-
-export const DefaultCategory = Object.freeze({
-  id: '0',
-  name: 'None',
-  subCategories: [],
-});
-
-export const CategoryInitialData = [
-  {
-    id: '0a5bf968-0497-4b15-96f5-51ca719204fa',
-    name: 'Food',
-    subCategories: [],
-  },
-  {
-    id: 'cf626e24-0bdf-47d4-9615-9e80b35d0acc',
-    name: 'Medical',
-    subCategories: [],
-  },
-  {
-    id: '4fbfd184-7f67-4ea0-a42b-6149317839d4',
-    name: 'Transportation',
-    subCategories: [],
-  },
-  {
-    id: 'c8963538-7a9a-4c61-a7d0-86d2dc6082d1',
-    name: 'Utilities',
-    subCategories: [],
-  },
-  {
-    id: 'c04cdb02-784c-4141-b295-cccca7c12c2e',
-    name: 'Housing',
-    subCategories: [],
-  },
-  {
-    id: '333a1aec-c514-426a-9ed2-1e8e6c313bfe',
-    name: 'Personal',
-    subCategories: [],
-  },
-  {
-    id: 'ae74b703-af3f-4a65-a996-942a5ed426a6',
-    name: 'Entertainment',
-    subCategories: [],
-  },
-  {
-    id: 'b018f0c0-41c0-41a9-bf87-b7b31918fb36',
-    name: 'Miscellaneous',
-    subCategories: [],
-  },
-];
