--- src/store/categories/index.js
@@ -1,4 +1,0 @@
-export * as actions from './actions';
-export * as constants from './constants';
-export * as selectors from './selectors';
-export { default as reducer } from './slice';
