--- src/store/categories/selectors.js
@@ -1,5 +1,0 @@
-export const selectCategories = (state) => state.categories;
-
-export const selectCategoryById = (state, categoryId) => {
-  return state.categories.find((category) => category.id === categoryId);
-};
