+++ src/store/constants.js
@@ -1,0 +1,12 @@
+export const AccountType = Object.freeze({
+  SAVINGS: 'Savings',
+  CHECKING: 'Checking',
+  CREDIT_CARD: 'Credit Card',
+});
+
+export const TransactionStatusEnum = Object.freeze({
+  COMPLETED: 'completed',
+  PENDING: 'pending',
+  SCHEDULED: 'scheduled',
+  PLANNED: 'planned',
+});
