+++ src/store/entities/slice.js
@@ -1,0 +1,1 @@
+import { createSlice } from '@reduxjs/toolkit';
