+++ src/store/repeatLinkTable/slice.js
@@ -1,0 +1,6 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const repeatLinkTable = createSlice({
+  name: 'repeatLinkTable',
+  reducers: {},
+});
