--- src/store/transactions/constants.js
@@ -1,15 +1,0 @@
-export const TransactionStatusEnum = Object.freeze({
-  PENDING: 'pending ',
-  COMPLETE: 'complete ',
-  SCHEDULED: 'scheduled ',
-  PLANNED: 'planned ',
-});
-
-export const TransactionFields = Object.freeze({
-  ID: 'id',
-  STATUS: 'status',
-  DATE: 'date',
-  AMOUNT: 'amount',
-  DESCRIPTION: 'description',
-  CATEGORY: 'categoryId',
-});
