--- src/store/transactions/index.js
@@ -1,4 +1,0 @@
-export * as actions from './actions';
-export * as constants from './constants';
-export { default as schemas } from './schemas';
-export * as reducers from './slice';
